chore[test]: floating point rounding issue

+++ src/models.py
@@ -1,4 +1,4 @@
-from decimal import ROUND_HALF_UP, Decimal
+from decimal import ROUND_CEILING, ROUND_HALF_UP, Decimal
 
 from pydantic import BaseModel, field_validator
 
@@ -26,9 +26,10 @@ class Account(BaseModel):
     @field_validator("amount", mode="before")
     @classmethod
     def round_float_to_int(cls, v):
-        if isinstance(v, float):
+        if isinstance(v, Decimal):
             return int(Decimal(str(v)).quantize(Decimal("1"), rounding=ROUND_HALF_UP))
         return v
+
 
 
 Memory.register_model("accounts", Account)
